Update docker-compose.yml for kong entrypoint
kong entrypoint corrupts kong.yml...might be due to lack of curly braces inside temp.yml.  symptom is resolved if kong.yml is removed and temp.yml is copied again w/ the variables. then the old entrypoint would do the replacement and start up.  sed here is an attempt to handle it when the stack is deployed w/o monkeying around w/ stuff.